Organize backend and generation preprocessing
