+++ Sources/PalmierPro/Backend/BackendStorage.swift
@@ -5,7 +5,7 @@ enum BackendStorage {
     static func uploadStaged(fileURL: URL, contentType: String) async throws -> String {
         let ticket = try await uploadTicket()
         guard let stagingURL = URL(string: ticket.uploadUrl) else {
-            throw GenerationBackendError.transport("Invalid staging URL")
+            throw BackendError.transport("Invalid staging URL")
         }
 
         var request = URLRequest(url: stagingURL)
@@ -19,25 +19,25 @@ enum BackendStorage {
     @MainActor
     private static func uploadTicket() async throws -> StagingTicket {
         guard let convex = AccountService.shared.convex else {
-            throw GenerationBackendError.notConfigured
+            throw BackendError.notConfigured
         }
         return try await convex.mutation("uploads:generateUploadTicket")
     }
 
     private static func assertHTTPOK(respData: Data, response: URLResponse) throws {
         guard let http = response as? HTTPURLResponse else {
-            throw GenerationBackendError.transport("Non-HTTP response")
+            throw BackendError.transport("Non-HTTP response")
         }
         if (200..<300).contains(http.statusCode) { return }
         let detail = String(data: respData, encoding: .utf8) ?? ""
         if let parsed = try? JSONDecoder().decode(BackendErrorEnvelope.self, from: respData) {
-            throw GenerationBackendError.api(
+            throw BackendError.api(
                 status: http.statusCode,
                 code: parsed.error.code,
                 message: parsed.error.message
             )
         }
-        throw GenerationBackendError.transport("HTTP \(http.statusCode): \(detail)")
+        throw BackendError.transport("HTTP \(http.statusCode): \(detail)")
     }
 }
