+++ Sources/PalmierPro/Transcription/TranscriptionBackend.swift
@@ -11,7 +11,7 @@ enum TranscriptionBackend {
         projectId: String?
     ) async throws -> BackendTranscriptionSubmit {
         guard let convex = AccountService.shared.convex else {
-            throw GenerationBackendError.notConfigured
+            throw BackendError.notConfigured
         }
         let args: [String: ConvexEncodable?] = [
             "storageId": storageId,
@@ -49,7 +49,7 @@ enum TranscriptionBackend {
     @MainActor
     private static func resultRef(jobId: String) async throws -> BackendTranscriptionResultRef {
         guard let convex = AccountService.shared.convex else {
-            throw GenerationBackendError.notConfigured
+            throw BackendError.notConfigured
         }
         return try await convex.action(
             "transcriptions:result",
@@ -60,7 +60,7 @@ enum TranscriptionBackend {
     @MainActor
     static func waitForResult(jobId: String) async throws -> TranscriptionResult {
         guard let publisher = subscribe(jobId: jobId) else {
-            throw GenerationBackendError.notConfigured
+            throw BackendError.notConfigured
         }
         for await job in jobStream(from: publisher) {
             guard let job else { continue }
